% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_abundances.R
\name{add_absolute_abundance}
\alias{add_absolute_abundance}
\title{Add absolute abundance to counts table}
\usage{
add_absolute_abundance(ta, spike_taxon, spike_added = spike_added)
}
\arguments{
\item{ta}{tidytacos object.}

\item{spike_taxon}{The taxon id of the spike.}

\item{spike_added}{The column name of the samples table which indicates how much spike was added per sample, e.g. 16S rRNA gene copy numbers added to the DNA extraction tube.}
}
\description{
\code{add_absolute_abundance} adds absolute abundance to the counts table of a
tidytacos object.
}
\details{
This function adds the absolute abundance per taxon and per sample to the counts table
of a tidytacos object under the variable name "absolute_abundance".
}
\examples{
# Initiate count matrix
x <- matrix(
  c(1500, 1300, 14, 280, 356, 9),
  ncol = 2
)
rownames(x) <- c("taxon1", "taxon2", "taxon3")
colnames(x) <- c("sample1", "sample2")

# Convert to tidytacos object
data <- create_tidytacos(x,
  taxa_are_columns = FALSE
)
data$samples$spike_added <- c(100, 150)

# Add total abundance
data <- data \%>\%
  add_absolute_abundance(spike_taxon = "t3")

}
