% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_samples.R
\name{add_alpha}
\alias{add_alpha}
\title{Add alpha diversity measures}
\usage{
add_alpha(ta, method = "invsimpson")
}
\arguments{
\item{ta}{A tidytacos object.}

\item{method}{The diversity measure to use, see \code{\link[vegan]{diversity}} for examples.}
}
\description{
\code{add_alpha} adds two alpha diversity measures to the sample table of a
tidytacos object.
}
\details{
This function adds two alpha diversity measures, observed richness and inverse
Simpson index, to the sample table of a tidytacos object under the variable
names "observed" and "inverse_simpson", respectively.
}
\examples{
# Initiate counts matrix
x <- matrix(
 c(1500, 1300, 280, 356),
 ncol = 2
)
rownames(x) <- c("taxon1", "taxon2")
colnames(x) <- c("sample1", "sample2")

# Convert to tidytacos object
data <- create_tidytacos(x,
                     taxa_are_columns = FALSE
                     )

# Add total abundance
data <- data \%>\%
 add_alpha()

}
