% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{add_clr_abundance}
\alias{add_clr_abundance}
\title{Perform a centered log ratio transformation on the readcounts.}
\usage{
add_clr_abundance(ta, overwrite = F)
}
\arguments{
\item{ta}{a tidytacos object}

\item{overwrite}{wether or not the counts table is to be overwritten with the transformed counts.}

\item{taxon}{unique identifier for the taxon, by default taxon_id.}

\item{sample}{unique identifier for the sample, by default sample_id.}

\item{counts}{variable name of the counts to be transformed in the counts table.}
}
\description{
Perform a centered log ratio transformation on the readcounts.
}
