% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codifab.R
\name{add_copca}
\alias{add_copca}
\title{Add compositional principal components to the sample table}
\usage{
add_copca(ta)
}
\arguments{
\item{ta}{A tidytacos object.}
}
\description{
\code{add_copca} performs a principal components analysis and 
adds the first two principal components to the sample table 
under column names "pca_1" and "pca_2".
}
\details{
Note that this function uses only the 50 most prevalant taxa 
unless \code{\link{add_logratio}} was executed with 
another value for 'max_taxa'.
}
