% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_samples.R
\name{add_dominant_taxa}
\alias{add_dominant_taxa}
\title{Adds the most dominant taxa in the sample to the sample table}
\usage{
add_dominant_taxa(ta, threshold_dominance = 0.5, taxon_name = taxon_id)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{threshold_dominance}{The relative abundance threshold for a taxon to be considered dominant.}

\item{taxon_name}{The column name of the taxa table that defines the taxon name.}
}
\value{
A tidytacos object with the dominant taxa added to the sample table and the Berger-Parker index of the most dominant taxon.
If the B-P index is lower than the threshold NA is returned.
}
\description{
\code{add_dominant_taxa()} adds the most dominant taxa in the sample
above a given relative abundance to the sample table.
Samples that do not have a dominant taxon will have NA in the dominant_taxon column.
}
\seealso{
Other sample-modifiers: 
\code{\link{add_alpha}()},
\code{\link{add_alphas}()},
\code{\link{add_metadata}()},
\code{\link{add_ord}()},
\code{\link{add_sample_clustered}()},
\code{\link{add_spike_ratio}()},
\code{\link{add_subsampled_alpha}()},
\code{\link{add_total_absolute_abundance}()},
\code{\link{add_total_count}()},
\code{\link{add_total_density}()},
\code{\link{cluster_samples}()}
}
\concept{sample-modifiers}
