% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codifab.R
\name{add_logratio}
\alias{add_logratio}
\title{Add logratios}
\usage{
add_logratio(ta, max_taxa = 50)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{max_taxa}{The maximum number of taxa to use.}
}
\value{
A tidytacos object with an extra table logratios
}
\description{
This function computes pairwise logratio values between all taxa and adds
these to the tidytacos object in the form of a table called logratios.
}
\details{
If `max_taxa` is smaller than the number of taxa in the dataset, the taxa with the highest
prevalence will be selected.

IMPORTANT: this function adds pseudocounts of one to all abundances before
calculating the logratios.
}
