% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_samples.R
\name{add_metadata}
\alias{add_metadata}
\title{Add metadata table to the tidytacos object}
\usage{
add_metadata(ta, metadata_tibble, table_type = "sample")
}
\arguments{
\item{ta}{tidytacos object.}

\item{table_type}{The type of table to add, either 'sample' or 'taxa'.}

\item{metadata}{A tibble containing sample data for each sample or taxon. Samples/taxa
should be rows, while metadata should be columns. At least one column
name needs to be shared with the sample or taxa tibble of ta. The default shared
column name is 'sample' for samples and 'taxon' for taxa.}
}
\description{
\code{add_metadata} adds a tibble to the tidytacos object.
}
\details{
This function adds a tibble containing metadata for each sample or taxon to the
tidytacos object. It is used after initiating a tidytacos object
using a numerical abundance matrix and the function
\code{\link{create_tidytacos}}.
}
\examples{
# Initiate counts matrix
x <- matrix(
 c(1500, 1300, 280, 356),
 ncol = 2
)
rownames(x) <- c("taxon1", "taxon2")
colnames(x) <- c("sample1", "sample2")

# Convert to tidytacos object
data <- create_tidytacos(x,
                     taxa_are_columns = FALSE
                     )

# Initiate sample tibble
sample <- c("sample1", "sample2")
environment <- c("food fermentation", "human stool")
sample_tibble <- tibble::tibble(sample, environment)

# Add sample tibble to tidytacos object
data <- data \%>\%
add_metadata(sample_tibble)

}
