% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_samples.R
\name{add_ord}
\alias{add_ord}
\title{Add ordination}
\usage{
add_ord(ta, distance = "bray", method = "pcoa", dims = 2, binary = FALSE, ...)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{distance}{The distance indices to use, see
\code{\link[vegan:vegdist]{vegan::vegdist()}}.}

\item{method}{The ordination method to use to calculate coordinates. Choices
are \code{pcoa}, \code{tsne}, \code{umap}.}

\item{dims}{The amount of dimensions to reduce the dissimilarities to.}

\item{binary}{Perform presence/absence standardisation before distance
computation or not.}

\item{...}{Additional arguments to pass to the ordination function:
either \code{\link[stats:cmdscale]{stats::cmdscale()}}, \code{\link[Rtsne:Rtsne]{Rtsne::Rtsne()}} or \code{\link[umap:umap]{umap::umap()}}.}
}
\value{
A tidytacos object with the ordination coordinates added.
}
\description{
\code{add_ord()} adds the first n dimensions of a dimensionality reduction
method performed on a given dissimilarity matrix as new variables to the
sample table of a tidytacos object.
}
\details{
This function calculates the dissimilarities between samples followed by
an ordination analysis. It will then add the first n dimensions to
the sample table of a tidytacos object named "ord1", "ord2", ... This
function will also add relative abundances if not present using
\code{\link[=add_rel_abundance]{add_rel_abundance()}}.
}
\examples{
# Initiate counts matrix
x <- matrix(
  c(1500, 1300, 280, 356, 456, 678),
  ncol = 3
)
rownames(x) <- c("taxon1", "taxon2")
colnames(x) <- c("sample1", "sample2", "sample3")

# Convert to tidytacos object
data <- create_tidytacos(x,
  taxa_are_columns = FALSE
)

# Add pcoa
data <- data \%>\%
  add_ord()

# The variances of the ordination dimensions can be accessed with
data$ord_variances
}
\seealso{
Other sample-modifiers: 
\code{\link{add_alpha}()},
\code{\link{add_alphas}()},
\code{\link{add_metadata}()},
\code{\link{add_sample_clustered}()},
\code{\link{add_spike_ratio}()},
\code{\link{add_total_absolute_abundance}()},
\code{\link{add_total_count}()},
\code{\link{add_total_density}()},
\code{\link{cluster_samples}()}

Other diversity-metrics: 
\code{\link{add_alpha}()},
\code{\link{add_alphas}()}
}
\concept{diversity-metrics}
\concept{sample-modifiers}
