% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_abundances.R
\name{add_rel_abundance}
\alias{add_rel_abundance}
\title{Add relative abundances to count table}
\usage{
add_rel_abundance(ta)
}
\arguments{
\item{ta}{A tidytacos object.}
}
\value{
A tidytacos object with relative abundances added to the count table.
}
\description{
\code{add_rel_abundance()} calculates relative abundances of taxa in samples
and adds them to the count table of a tidytacos object under the column name
"rel_abundance".
}
\examples{
# Initiate count matrix
x <- matrix(
  c(1500, 1300, 280, 356),
  ncol = 2
)
rownames(x) <- c("taxon1", "taxon2")
colnames(x) <- c("sample1", "sample2")

# Convert to tidytacos object
data <- create_tidytacos(x,
  taxa_are_columns = FALSE
)

# Add relative abundance
data <- data \%>\% add_rel_abundance()
}
\seealso{
Other count-modifiers: 
\code{\link{add_absolute_abundance}()},
\code{\link{add_density}()}
}
\concept{count-modifiers}
