% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_samples.R
\name{add_spike_ratio}
\alias{add_spike_ratio}
\title{Add spike ratio}
\usage{
add_spike_ratio(ta, spike_taxon)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{spike_taxon}{The taxon_id of the spike.}
}
\value{
A tidytacos object with the spike ratio added to the sample table.
}
\description{
\code{add_spike_ratio()} calculates the ratio of non-spike to spike reads for
each sample and adds this to the sample table under the name "spike_ratio".
}
\details{
This function is useful if a DNA spike was added prior to sequencing and is
based on the method described by
\href{https://doi.org/10.1016/j.soilbio.2016.02.003}{Smets et al., 2016}.

Without calculating absolute abundances, the spike ratio allows to compare absolute abundances between sample. For example, if the spike ration of one sample is twice that of another, then the absolute number of sequenced strands at the time of spiking in the one sample is twice that of the other sample.
}
\examples{
# Initiate counts matrix
x <- matrix(
  c(1500, 1300, 280, 356),
  ncol = 2
)
rownames(x) <- c("taxon1", "taxon2")
colnames(x) <- c("sample1", "sample2")

# Convert to tidytacos object
data <- create_tidytacos(x,
  taxa_are_columns = FALSE
)

# Add total abundance
data <- data \%>\%
  add_spike_ratio(spike_taxon = "t1")
}
\seealso{
Other sample-modifiers: 
\code{\link{add_alpha}()},
\code{\link{add_alphas}()},
\code{\link{add_metadata}()},
\code{\link{add_ord}()},
\code{\link{add_sample_clustered}()},
\code{\link{add_total_absolute_abundance}()},
\code{\link{add_total_count}()},
\code{\link{add_total_density}()},
\code{\link{cluster_samples}()}
}
\concept{sample-modifiers}
