% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_samples.R
\name{add_subsampled_alpha}
\alias{add_subsampled_alpha}
\title{Add alpha diversity measures using subsampling}
\usage{
add_subsampled_alpha(
  ta,
  min_lib_size = NULL,
  method = "shannon",
  itterations = 100
)
}
\arguments{
\item{ta}{a tidytacos object.}

\item{min_lib_size}{the minimum lib size samples need to have.
Samples with lower lib sizes will be discarded and samples with a higher readcount will be itteratively
subsampled to this readcount to allow for a fair comparison across read_depths.}

\item{method}{The diversity measure to use,
see \code{\link[vegan:diversity]{vegan::diversity()}} for further information on these.}

\item{itterations}{the amount of itterations for subsampling. Please report this number in your research.}
}
\value{
A tidytacos object with the selected alpha diversity measure added.
}
\description{
\code{add_subsampled_alpha()} adds selected alpha diversity measures to the sample table of a
tidytacos object using an itterative subsampling process.
}
\seealso{
Other sample-modifiers: 
\code{\link{add_alpha}()},
\code{\link{add_alphas}()},
\code{\link{add_dominant_taxa}()},
\code{\link{add_metadata}()},
\code{\link{add_ord}()},
\code{\link{add_sample_clustered}()},
\code{\link{add_spike_ratio}()},
\code{\link{add_total_absolute_abundance}()},
\code{\link{add_total_count}()},
\code{\link{add_total_density}()},
\code{\link{cluster_samples}()}

Other diversity-metrics: 
\code{\link{add_alpha}()},
\code{\link{add_alphas}()},
\code{\link{add_ord}()}
}
\concept{diversity-metrics}
\concept{sample-modifiers}
