% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_taxa.R
\name{add_taxon_name}
\alias{add_taxon_name}
\title{Add sensible taxon name to taxon table}
\usage{
add_taxon_name(ta, method = "mean_rel_abundance", include_species = FALSE)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{method}{The method by which to arrange the taxon names. Currently only
mean_rel_abundance.}

\item{include_species}{Whether to include the species name or not.}
}
\value{
A tidytacos object.
}
\description{
\code{add_taxon_name()} creates sensible taxa names by
-under default conditions- combining the genus name with a number.
The number is only added if there is more than one taxon of that genus.
The number indicates the rank of abundance, with 1 indicating
the taxon with the highest mean relative abundance within the genus.
If genus classification is not available
the next most detailed taxonomic rank which is available is used.
The sensible taxon name is added to the taxon table
under the column name \code{taxon_name}.
}
\examples{
urt_g <- urt \%>\% add_taxon_name()
# add the species name if present (which is often uncertain in amplicon data)
urt_s <- urt \%>\% add_taxon_name(include_species = TRUE)
}
\seealso{
Other taxa-modifiers: 
\code{\link{add_mean_rel_abundance}()},
\code{\link{add_prevalence}()},
\code{\link{add_taxon_name_color}()}
}
\concept{taxa-modifiers}
