% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_samples.R
\name{add_total_count}
\alias{add_total_count}
\title{Add total read count per sample}
\usage{
add_total_count(ta)
}
\arguments{
\item{ta}{A tidytacos object.}
}
\value{
A tidytacos object with the total read count per sample added.

A tidytacos object with the total read count per sample added.
}
\description{
\code{add_total_count()} adds the total read count per sample to the sample
table of a tidytacos object under the variable name total_count.
}
\examples{
# Initiate counts matrix
x <- matrix(
  c(1500, 1300, 280, 356),
  ncol = 2
)
rownames(x) <- c("taxon1", "taxon2")
colnames(x) <- c("sample1", "sample2")

# Convert to tidytacos object
data <- create_tidytacos(x,
  taxa_are_columns = FALSE
)

# Add total counts
data <- data \%>\%
  add_total_count()
}
\seealso{
Other sample-modifiers: 
\code{\link{add_alpha}()},
\code{\link{add_alphas}()},
\code{\link{add_metadata}()},
\code{\link{add_ord}()},
\code{\link{add_sample_clustered}()},
\code{\link{add_spike_ratio}()},
\code{\link{add_total_absolute_abundance}()},
\code{\link{add_total_density}()},
\code{\link{cluster_samples}()}
}
\concept{sample-modifiers}
