% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_samples.R
\name{add_total_density}
\alias{add_total_density}
\title{Add total densities of samples}
\usage{
add_total_density(
  ta,
  spike_taxon,
  spike_added = spike_added,
  material_sampled = material_sampled
)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{spike_taxon}{The taxon id of the spike.}

\item{spike_added}{The column name of the samples table which indicates how
much spike was added per sample, e.g. 16S rRNA gene copy numbers added to
the DNA extraction tube.}

\item{material_sampled}{The column name indicating the amount of material
from which DNA was extracted, e.g gram of soil. This parameter encourages
researchers to consider that absolute abundances are only meaningful if
they can be translated into densities.}
}
\value{
A tidytacos object with the total
densities added to the sample table.
}
\description{
\code{add_total_density()} adds the total microbial density to the sample table
of a tidytacos object under the column name "total_density".
}
\examples{
# Initiate count matrix
x <- matrix(
  c(1500, 1300, 14, 280, 356, 9),
  ncol = 2
)
rownames(x) <- c("taxon1", "taxon2", "taxon3")
colnames(x) <- c("sample1", "sample2")

# Convert to tidytacos object
data <- create_tidytacos(x,
  taxa_are_columns = FALSE
)
data$samples$spike_added <- c(100, 150)
data$samples$material_sampled <- c(1, 5)

# Add total abundance
data <- data \%>\%
  add_total_density(spike_taxon = "t3")
}
\seealso{
Other sample-modifiers: 
\code{\link{add_alpha}()},
\code{\link{add_alphas}()},
\code{\link{add_metadata}()},
\code{\link{add_ord}()},
\code{\link{add_sample_clustered}()},
\code{\link{add_spike_ratio}()},
\code{\link{add_total_absolute_abundance}()},
\code{\link{add_total_count}()},
\code{\link{cluster_samples}()}
}
\concept{sample-modifiers}
