% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_models.R
\name{align_lda_topics}
\alias{align_lda_topics}
\title{Aligns LDA topics across models}
\usage{
align_lda_topics(
  ta,
  k_range = 2:10,
  seed = 42,
  min_prevalence = 0.05,
  taxon = taxon_id,
  sample = sample_id,
  ...
)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{k_range}{A range of topic numbers to estimate.}

\item{seed}{A seed for reproducibility.}

\item{min_prevalence}{The lowest percentage (0-1) of samples
taxa need to be present in for the taxa to be used in model estimation.}

\item{taxon}{The column name in the taxa table with taxa identifiers.}

\item{sample}{The column name in the sample table with sample identifiers.}

\item{...}{
  Arguments passed on to \code{\link[alto:run_lda_models]{alto::run_lda_models}}
  \describe{
    \item{\code{data}}{(required) a \code{matrix}, \code{data.frame} or
\code{slam::simple_triplet_matrix} containing the counts (integers) of each
feature (e.g. words) and each sample (or document). If data is provided as
\code{matrix} or \code{data.frame}, each row is a sample, each column is a
feature.}
    \item{\code{lda_varying_params_lists}}{(required) a \code{list} specifying the
parameter for each models that needs to be ran. Currently, supported
parameters are "k" (the number of topic), "method" ("VEM" or "Gibbs"), and
"control", a list of type \code{LDAcontrol}. See \code{topicmodels::LDA} for
details and below for examples.}
    \item{\code{lda_fixed_params_list}}{(optional) a \code{list} specifying the
parameters common to all models to be fitted. Values provided by
\code{lda_fixed_params_list} are overwritten by those provided by
\code{lda_varying_params_lists}.}
    \item{\code{dir}}{(optional) a \code{character} specifying the directory in which
individual LDA models should be stored. If not specified, individual LDA
models are not stored. This option is especially useful for data exploration
as it allows to save execution time if one wishes to add models to an
existing model list. (see examples)}
    \item{\code{reset}}{(optional, default = \code{FALSE}). Should any cached models in
the save directory be cleared?}
    \item{\code{verbose}}{(optional, default = \code{FALSE}) Print verbose output while
running models?}
  }}
}
\value{
A list of estimated models, aligned topics using the product method
and aligned topics using the transport method.
}
\description{
\code{align_lda_topics()} estimates LDA models for a range of topic numbers
and aligns the topics across models using the product and transport method.
}
