% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{as_phyloseq}
\alias{as_phyloseq}
\title{Convert tidytacos object to phyloseq object}
\usage{
as_phyloseq(ta, sample = sample, taxon = taxon_id)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{sample}{The sample names required for a phyloseq object. Default is
"sample" column of the sample table of the tidytacos object.}

\item{taxon}{The taxon names required for a phyloseq object. Default is the
"taxon_id" column in the taxon table of the tidytacos object.}
}
\description{
\code{as_phyloseq} returns a phyloseq object given a tidytacos object.
}
\details{
This function will convert a tidytacos object into a phyloseq object for
alternative processing using the phyloseq package. To convert from a phyloseq
object to a tidytacos object use \code{\link{as_tidytacos}}.
}
