% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_models.R
\name{calculate_lda_perplexities}
\alias{calculate_lda_perplexities}
\title{Calculate LDA model perplexities for a range of topic numbers}
\usage{
calculate_lda_perplexities(ta, k_range = 2:10, seed = 42, ...)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{k_range}{A range of topic numbers to estimate.}

\item{seed}{A seed for reproducibility.}

\item{...}{
  Arguments passed on to \code{\link[=perform_lda]{perform_lda}}
  \describe{
    \item{\code{k}}{The number of topics to estimate.}
    \item{\code{min_prevalence}}{The lowest percentage (0-1) of samples
taxa need to be present in for the taxa to be used in model estimation.}
    \item{\code{taxon}}{The column name in the taxa table with taxa identifiers.}
    \item{\code{sample}}{The column name in the sample table with sample identifiers.}
  }}
}
\value{
A tibble with the perplexity of each model and the number of topics.
}
\description{
\code{calculate_lda_perplexities()} estimates LDA models for a range of topic numbers
and returns the perplexity of each model.
}
