% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_samples.R
\name{cluster_samples}
\alias{cluster_samples}
\title{Clusters samples in desired number of clusters and adds these to the sample table.}
\usage{
cluster_samples(ta, n_clusters)
}
\arguments{
\item{ta}{tidytacos object.}

\item{n_clusters}{Numerical. Number of desired clusters.}
}
\description{
\code{cluster_samples} adds a new variable to the samples tibble of a
tidytacos object defining to what cluster a sample belongs.
}
\details{
This function calculates the Bray-Curtis distance between samples followed by
hierarchical average linkage clustering of samples. The user provides a
number of desired clusters which will be used to assign the samples to. A new
variable named "cluster" will be added to the samples tibble of a
tidytacos object defining to what cluster a sample belongs.
}
\examples{
# Initiate count matrix
x <- matrix(
 c(1500, 1300, 280, 356),
 ncol = 2
)
rownames(x) <- c("taxon1", "taxon2")
colnames(x) <- c("sample1", "sample2")

# Convert to tidytacos object
data <- create_tidytacos(x,
                     taxa_are_columns = FALSE
                     )

# Add total abundance
data <- data \%>\%
 cluster_samples(n_clusters = 2)

}
