% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{cluster_sparcc}
\alias{cluster_sparcc}
\title{Performs Markov Clustering on a sparcc network matrix generated by \code{\link{filter_sparcc}}. 
and returns a tibble with clusters and taxon identifiers. Optionally, the network can be visualized.}
\usage{
cluster_sparcc(sparcc, min_n = 3, visualize = F)
}
\arguments{
\item{min_n}{minimum number of taxa per cluster, smaller clusters are filtered out.}

\item{visualize}{wether to visualize the network clusters or not.}

\item{network}{sparcc network in matrix form.}
}
\description{
Performs Markov Clustering on a sparcc network matrix generated by \code{\link{filter_sparcc}}. 
and returns a tibble with clusters and taxon identifiers. Optionally, the network can be visualized.
}
