% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{cluster_taxa}
\alias{cluster_taxa}
\title{Performs SparCC network analysis on a tidytacos object 
and then performs Markov Clustering on the network to 
annotate taxa of the largest clusters in the tidytacos object.}
\usage{
cluster_taxa(
  ta,
  min_occurrence = 0.05,
  network_thresh = 0.1,
  min_n_cluster = 3,
  taxon_name = taxon,
  sample_name = sample
)
}
\arguments{
\item{ta}{a tidytacos object.}

\item{min_occurrence}{Percentage of samples the taxon needs to be present in for it to be considered in the analysis.}

\item{network_thresh}{absolute value of correlations below this threshold are 
filtered out.}

\item{min_n_cluster}{minimum number of taxa per cluster, smaller clusters are filtered out.}

\item{taxon_name}{unique name to use for the taxa, by default taxon_id is used.}

\item{sample_name}{unique name to use for the samples, by default sample_id is used.}
}
\description{
Performs SparCC network analysis on a tidytacos object 
and then performs Markov Clustering on the network to 
annotate taxa of the largest clusters in the tidytacos object.
}
