% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{counts_matrix}
\alias{counts_matrix}
\title{Return a counts matrix}
\usage{
counts_matrix(
  ta,
  sample_name = sample_id,
  taxon_name = taxon_id,
  value = count,
  keep_empty_samples = FALSE
)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{sample_name}{The name of the variable in the sample table to use as row
names (unquoted).}

\item{taxon_name}{The name of the variable in the taxon table to use as
column names (unquoted).}

\item{value}{The name of the variable in the counts table to use as count}

\item{keep_empty_samples}{Should empty samples be included in the matrix?}
}
\value{
A matrix with count values.
}
\description{
This function returns a matrix with taxon counts; the rows are samples and
the columns are taxa.
}
