% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{counts_tidy}
\alias{counts_tidy}
\title{Convert matrix with counts to tidy data frame}
\usage{
counts_tidy(counts_matrix, taxa_are_columns = TRUE, value = "counts")
}
\arguments{
\item{counts_matrix}{The count matrix that will be converted.}

\item{taxa_are_columns}{A logical scalar. Are the taxa defined in columns?
Default is TRUE.}

\item{value}{Name of resulting colum containing the count data. Default
is "counts".}
}
\description{
`counts_tidy()` returns a tidy data frame given a numerical counts
matrix.
}
\details{
This function will convert a numerical counts matrix into a tidy data
frame. To convert a tidy data frame into a numerical counts matrix
use ´[counts_matrix()].
}
