% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{create_tidytacos}
\alias{create_tidytacos}
\title{Initiate tidytacos object}
\usage{
create_tidytacos(counts_matrix, taxa_are_columns = TRUE)
}
\arguments{
\item{counts_matrix}{Numerical matrix containing the count data.}

\item{taxa_are_columns}{A logical scalar. Are the taxa defined in columns?}
}
\value{
A tidytacos object.
}
\description{
\code{create_tidytacos()} returns a tidytacos object given a numeric matrix.
}
\details{
This function initiates a tidytacos object based on a numeric matrix. It
will automatically create a dummy taxa table and sample table which will need to be
updated using the function \code{\link[=add_metadata]{add_metadata()}}.
When the taxa table is updated, the rank names can be set using the function \code{\link[=set_rank_names]{set_rank_names()}}
to make the tidytacos object aware of the taxonomy levels.
}
\examples{
# Initiate count matrix
x <- matrix(
  c(1500, 1300, 280, 356),
  ncol = 2
)
rownames(x) <- c("taxon1", "taxon2")
colnames(x) <- c("sample1", "sample2")

# Convert to tidytacos object
data <- create_tidytacos(x,
  taxa_are_columns = FALSE
)
# Add taxonomy information
taxonomy <- tibble::tibble(
  taxon = c("taxon1", "taxon2"),
  domain = c("Bacteria", "Bacteria"),
  phylum = c("Bacillota", "Pseudomonadota"),
  class = c("Bacilli", "Gammaproteobacteria"),
  order = c("Lactobacillales", "Enterobacteriales"),
  family = c("Lactobacillaceae", "Enterobacteriaceae"),
  genus = c("Lactobacillus", "Escherichia"),
  species = c("Lactobacillus crispatus", "Escherichia coli")
)

data <- add_metadata(data, taxonomy, table_type = "taxa")
# rank names are inferred from the taxa table, but can be set manually
# if we're not happy with the inferred rank names.
data <- set_rank_names(data, c("domain", "phylum", "class", "order", "family", "genus"))
}
\seealso{
Other import-methods: 
\code{\link{from_dada}()},
\code{\link{from_phyloseq}()},
\code{\link{read_tidytacos}()}
}
\concept{import-methods}
