% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{filter_counts}
\alias{filter_counts}
\title{Filter the counts}
\usage{
filter_counts(ta, ...)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{...}{Filter criteria for the counts table.}
}
\description{
Filter the counts
}
