% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{filter_samples}
\alias{filter_samples}
\title{Filter the samples}
\usage{
filter_samples(ta, ...)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{...}{Filter criteria for the samples table.}
}
\value{
A tidytacos object.
}
\description{
Filter the samples
}
\examples{

# subset urt to keep only nasopharynx samples
urt_nf <- urt \%>\% filter_samples(location == "NF")
# subset urt to keep only samples from plate 1 and 2
urt_plate_1_2 <- urt \%>\% filter_samples(plate \%in\% c(1, 2))
# subset the blanks in leaf
leaf_blanks <- leaf \%>\% filter_samples(startsWith(description, "BLANK"))

}
