% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{filter_sparcc}
\alias{filter_sparcc}
\title{Filters the output of \code{\link{sparcc}} to a minimal treshold 
and transforms to matrixfor downstream clustering or heatplot visualization.}
\usage{
filter_sparcc(sparcc, treshold = 0.1)
}
\arguments{
\item{treshold}{absolute value of correlations below this threshold are 
filtered out.}

\item{ta}{a tidytacos object}
}
\description{
Filters the output of \code{\link{sparcc}} to a minimal treshold 
and transforms to matrixfor downstream clustering or heatplot visualization.
}
