% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{from_dada}
\alias{from_dada}
\title{DADA2 to a tidytacos object}
\usage{
from_dada(seqtab, taxa, taxa_are_columns = TRUE)
}
\arguments{
\item{seqtab}{Sequence table, output of \code{\link[dada2:makeSequenceTable]{dada2::makeSequenceTable()}}.}

\item{taxa}{Taxa table, output of \code{\link[dada2:assignTaxonomy]{dada2::assignTaxonomy()}}.}

\item{taxa_are_columns}{A logical scalar. Are the taxa defined in columns?}
}
\value{
A tidytacos object.
}
\description{
\code{\link[=from_dada]{from_dada()}} returns a tidytacos object given a seqtab and taxa object from dada2.
}
\details{
This function will convert two dada2 objects or files into a tidytacos object.
}
\examples{
seqtab <- system.file("extdata", "dada2", "seqtab.rds", package = "tidytacos")
taxa <- system.file("extdata", "dada2", "taxa.rds", package = "tidytacos")
taco <- from_dada(seqtab, taxa)
}
\seealso{
Other import-methods: 
\code{\link{create_tidytacos}()},
\code{\link{from_phyloseq}()},
\code{\link{read_tidytacos}()}
}
\concept{import-methods}
