% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{from_phyloseq}
\alias{from_phyloseq}
\title{Convert phyloseq object to tidytacos object}
\usage{
from_phyloseq(ps)
}
\arguments{
\item{ps}{Phyloseq object.}
}
\description{
`from_phyloseq()` returns a tidytacos object given a phyloseq
object.
}
\details{
This function will convert a phyloseq object into a tidytacos object. To
convert from a tidytacos object to a phyloseq object use
[as_phyloseq()].
}
