% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{group_samples}
\alias{group_samples}
\title{Group the samples}
\usage{
group_samples(ta, ...)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{...}{Grouping criteria for the samples table.}
}
\value{
A (named) list of tidytacos object.
}
\description{
Group the samples
}
\examples{

urt_groups <- urt \%>\% group_samples(location)
# subset urt to keep only nasopharynx samples
urt_nf <- urt_groups$NF
}
