% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytacos-package.R
\docType{class}
\name{grouped_taco}
\alias{grouped_taco}
\title{An S4 class to store a grouped tidytaco with \link{group_samples}}
\description{
This S4 class stores the output from \link{group_samples}.
}
\section{Slots}{

\describe{
\item{\code{groups}}{the different grouped tidytacos.}
}}

