% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_models.R
\name{ldaplot_beta}
\alias{ldaplot_beta}
\title{Plot LDA beta values}
\usage{
ldaplot_beta(lda_alignment, show_product = TRUE, ...)
}
\arguments{
\item{lda_alignment}{The output of \code{align_lda_topics()}.}

\item{show_product}{Whether to show the product or transport alignment.}

\item{...}{
  Arguments passed on to \code{\link[alto:plot_beta]{alto::plot_beta}}
  \describe{
    \item{\code{x}}{(required) An alignment class object resulting from
\code{align_topics}.}
    \item{\code{models}}{Which models to display in the heatmap? Defaults to
\code{"all"}, meaning that all models are shown. If given \code{"last"}, only
the last model in the models list will be plotted. If given a vector of
characters, it will plot only models whose names in the original models list
match. Similarly, if given a list of integers, only the models lying at those
indices in the original model list will be visualized.}
    \item{\code{filter_by}}{(optional, default = \code{"beta"}) a character specifying
if the data (beta matrices) should be filtered by the average \code{"beta"}
across topics or by the \code{"distinctiveness"} of the features.}
    \item{\code{x_axis}}{(optional, default = \code{"index"}) a character specifying
if the x-axis should display topic indices (\code{"index"}) such that they
match the alignment plot order or topic names (\code{"label"}).}
    \item{\code{threshold}}{(optional, default = 0.001)
Words (features) with less than this average beta or
distinctiveness across all topics are ignored}
    \item{\code{n_features}}{(optional) alternative to \code{threshold}. The maximum
number of words (features) to display along rows of the plot.}
    \item{\code{beta_aes}}{Should word probabilities within a topic be encoded using
circle size (\code{"size"}) or opacity (\code{"alpha"}) ? Defaults to
\code{"size"}.}
    \item{\code{color_by}}{(optional) What should the color of topics and weights
encode? Defaults to 'path'. Other possible arguments are 'coherence',
'refinement', or 'topic'.}
  }}
}
\value{
A ggplot object.
}
\description{
\code{ldaplot_beta()} plots the beta values of LDA topics.
}
