% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{mutate_samples}
\alias{mutate_samples}
\title{Create extra variables in the sample table}
\usage{
mutate_samples(ta, ...)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{...}{Mutate criteria for the samples table.}
}
\value{
A tidytacos object.
}
\description{
Create extra variables in the sample table
}
\examples{

# change the sample column to lowercase
urt <- urt \%>\% mutate_samples(sample = tolower(sample))

}
