% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{mutate_taxa}
\alias{mutate_taxa}
\title{Create extra variables in the taxa table}
\usage{
mutate_taxa(ta, ...)
}
\arguments{
\item{ta}{A tidytacos object.}
}
\description{
Create extra variables in the taxa table
}
