% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{network}
\alias{network}
\title{Perform network inference with SparCC on a tidytacos object, 
after rarefaction of the taxa. See \link[SpiecEasi]{sparcc}.}
\usage{
network(
  ta,
  min_occurrence = 0.05,
  taxon_name = taxon,
  sample_name = sample,
  ...
)
}
\arguments{
\item{ta}{a tidytacos object}

\item{min_occurrence}{Percentage of samples the taxon needs to be present in for it to be considered in the analysis.}

\item{iter}{Number of iterations in the outer loop}

\item{inner_iter}{Number of iterations in the inner loop}

\item{th}{absolute value of correlations below this threshold are 
considered zero by the inner SparCC loop.}
}
\description{
Perform network inference with SparCC on a tidytacos object, 
after rarefaction of the taxa. See \link[SpiecEasi]{sparcc}.
}
