% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_samples.R
\name{perform_anosim}
\alias{perform_anosim}
\title{Perform anosim test}
\usage{
perform_anosim(ta, group, ...)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{group}{A column in the sample table to group the samples on.}

\item{distance}{The dissimilarity measure to use.}

\item{permutations}{The number of permutations.}
}
\description{
\code{perform_anosim} performs the anosim test for statistical difference
between groups of samples. The null hypothesis is that there is no difference
between microbial communities in the groups of samples.
}
