% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{rarefy}
\alias{rarefy}
\title{Rarefy the samples to a given number of reads}
\usage{
rarefy(ta, n, replace = F)
}
\arguments{
\item{n}{Subsample size for rarefying the community.}

\item{replace}{Whether to replace the read after it has been selected for the subsample so it can be sampled again? Default is FALSE.}
}
\description{
This function performs rarefying. Make sure that all samples contain at least
the minimum number of reads; otherwise, an error might be thrown.
}
