% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{reset_ids}
\alias{reset_ids}
\title{Reset the taxon and sample IDs}
\usage{
reset_ids(ta, keep_prev = FALSE)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{keep_prev}{A logical scalar.
Should the previous IDs be kept in a column called taxon_id_prev?}
}
\description{
Reset the taxon and sample IDs
}
