% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{sample_taco}
\alias{sample_taco}
\title{Take a sample of a tidytacos object}
\usage{
sample_taco(
  tt,
  n,
  group_col,
  taxon_identifier = sequence,
  replace = FALSE,
  ...
)
}
\arguments{
\item{tt}{A tidytacos object.}

\item{n}{The amount of samples that need to be returned.}

\item{group_col}{An optional name of a field in the sample table
which the samples need to be evenly distributed over.}

\item{taxon_identifier}{The column that uniquely identifies a taxa}

\item{replace}{Replace selected samples so they can be picked again in sampling.}

\item{...}{
  Arguments passed on to \code{\link[base:sample]{base::sample}}
  \describe{
    \item{\code{x}}{either a vector of one or more elements from which to choose,
    or a positive integer.  See \sQuote{Details.}}
    \item{\code{size}}{a non-negative integer giving the number of items to choose.}
    \item{\code{prob}}{a vector of probability weights for obtaining the elements
    of the vector being sampled.}
  }}
}
\value{
A tidytacos object.
}
\description{
\code{sample_taco()} takes a random subset of samples from the original tidytacos object.
}
