% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{select_counts}
\alias{select_counts}
\title{Retain or remove a set of count variables}
\usage{
select_counts(ta, ...)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{...}{Selection criteria for the counts table.}
}
\value{
A tidytacos object.
}
\description{
Retain or remove a set of count variables
}
\examples{

# add a column to the counts table
leaf_ab <- leaf \%>\% add_rel_abundance()
# remove that column again
leaf_ab \%>\% select_counts(-rel_abundance)
}
