% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{select_samples}
\alias{select_samples}
\title{Retain or remove a set of sample variables}
\usage{
select_samples(ta, ...)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{...}{Selection criteria for the samples table.}
}
\value{
A tidytacos object.
}
\description{
Retain or remove a set of sample variables
}
\examples{
# remove the condition column from the samples table
urt \%>\% select_samples(-condition)
# keep only the sample_id, location and method columns
urt \%>\% select_samples(sample_id, location, method)
}
