% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_abundances.R
\name{srs_normalize}
\alias{srs_normalize}
\title{Normalize counts using Scaling with Ranked Subsampling (SRS)
\code{srs_normalize()} uses the SRS method to equalize sampling depth to
a chosen value.}
\usage{
srs_normalize(ta, target_reads)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{target_reads}{the sequencing depth to which samples are subsampled. Samples below this threshold are discarded.}
}
\value{
A tidytacos object where the reads are subsampled to the chosen value using SRS.
}
\description{
Normalize counts using Scaling with Ranked Subsampling (SRS)
\code{srs_normalize()} uses the SRS method to equalize sampling depth to
a chosen value.
}
\seealso{
Other count-modifiers: 
\code{\link{add_absolute_abundance}()},
\code{\link{add_density}()},
\code{\link{add_rel_abundance}()}
}
\concept{count-modifiers}
