% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codifab.R
\name{tacoplot_codifab}
\alias{tacoplot_codifab}
\title{Generate a compositional differential abundance plot}
\usage{
tacoplot_codifab(ta, diffabun_var)
}
\arguments{
\item{ta}{A tidytacos object}

\item{diffabun_var}{The variable with differential abundances in the
taxon_pair table}
}
\value{
A ggplot object
}
\description{
This function returns a plot to visualize differential abundance of taxa
between conditions, compared to all other taxa as references. These
differential abundances should already have been calculated with
[add_codifab].
}
\details{
Significance of tests is determined by capping the false discovery rate at
10%, using the method of Benjamini and Yekutieli, which is developed for
non-independent tests. See [p.adjust].
}
