% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{tacoplot_ord_ly}
\alias{tacoplot_ord_ly}
\title{Return an interactive pcoa plot of the samples}
\usage{
tacoplot_ord_ly(
  ta,
  x = NULL,
  samplenames = sample_id,
  ord = "pcoa",
  dims = 2,
  distance = "bray",
  stat.method = "mantel",
  palette = NULL,
  title = NULL,
  ...
)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{x}{A string, representing the column name used to color the sample
groups on.}

\item{samplenames}{the column in the sample table with the samplenames, defaults to sample_id.}

\item{ord}{the ordination technique to use. Choice from pcoa, tsne and umap.}

\item{dims}{the amount of dimensions to plot, 2 or 3.}

\item{distance}{the distance algorithm to use, see \code{\link[vegan]{vegdist}}.}

\item{stat.method}{the statistic to print on the figure, choice from mantel and anosim.}

\item{palette}{A vector of colors, used as the palette for coloring sample}

\item{title}{a string to display as title of the plot.
groups.}
}
\description{
Return an interactive pcoa plot of the samples
}
