% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{tacoplot_prevalences}
\alias{tacoplot_prevalences}
\title{Return a heatmap of all taxa above a certain threshold prevalence per condition, clusters them and compares prevalences with a fisher_test.}
\usage{
tacoplot_prevalences(
  ta,
  condition,
  cutoff = 0.1,
  fisher = T,
  adjp_method = "fdr",
  ...
)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{condition}{The row name of the condition which rrevalences are to be compared.}

\item{cutoff}{The minimum prevalence of a taxon to be included in the heatmap.}

\item{fisher}{Run a fisher test on the relative prevalences in each condition 
and plot the resulting adjusted p-values as *(<.05), **(<.01), ***(<.001) or ****(<.0001).}

\item{adjp_method}{The method to adjust the p-values, see \code{\link[rstatix]{adjust_pvalue}}.}
}
\description{
See pheatmap for additional arguments that can be given for this heatmap.
Example usage: tacoplot_prevalences(urt,location,treeheight_row=0, cutree_rows=4,fontsize=6,cellwidth=15)
}
