% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{tacoplot_scree}
\alias{tacoplot_scree}
\title{Return a scree plot to visualize the eigenvalues of the PCA.}
\usage{
tacoplot_scree(ta, ...)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{...}{
  Arguments passed on to \code{\link[factoextra:eigenvalue]{factoextra::fviz_eig}}
  \describe{
    \item{\code{X}}{an object of class PCA, CA, MCA, FAMD, MFA and HMFA [FactoMineR]; prcomp 
and princomp [stats]; dudi, pca, coa and acm [ade4]; ca and mjca [ca 
package].}
    \item{\code{choice}}{a text specifying the data to be plotted. Allowed values are 
"variance" or "eigenvalue".}
    \item{\code{geom}}{a text specifying the geometry to be used for the graph. Allowed 
values are "bar" for barplot, "line" for lineplot or c("bar", "line") to use
both types.}
    \item{\code{barfill}}{fill color for bar plot.}
    \item{\code{barcolor}}{outline color for bar plot.}
    \item{\code{linecolor}}{color for line plot (when geom contains "line").}
    \item{\code{ncp}}{a numeric value specifying the number of dimensions to be shown.}
    \item{\code{addlabels}}{logical value. If TRUE, labels are added at the top of bars 
or points showing the information retained by each dimension.}
    \item{\code{hjust}}{horizontal adjustment of the labels.}
    \item{\code{main,xlab,ylab}}{plot main and axis titles.}
    \item{\code{ggtheme}}{function, ggplot2 theme name. Default value is theme_pubr().
Allowed values include ggplot2 official themes: theme_gray(), theme_bw(),
theme_minimal(), theme_classic(), theme_void(), ....}
  }}
}
\description{
Return a scree plot to visualize the eigenvalues of the PCA.
}
\examples{
urt \%>\% tacoplot_scree()
}
