% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{tacoplot_stack_ly}
\alias{tacoplot_stack_ly}
\title{Return an interactive bar plot of the samples}
\usage{
tacoplot_stack_ly(ta, n = 12, x = sample_clustered, ...)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{n}{An integer, representing the amount of colors used to depict
different taxa.}

\item{x}{A string, representing the column name used to label the x-axis}

\item{order_by}{an optional column name to order the samples by. 
For examples order_by=sample would order the x-axis by the sample names instead of by similar profiles.}
}
\description{
Plots an interactive stacked bar plot of the samples in the tidytacos object to inspect the taxonomic profile.
}
