% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{tacoplot_zoom}
\alias{tacoplot_zoom}
\title{Return a visualization designed for a small number of samples}
\usage{
tacoplot_zoom(ta, sample = sample_id, n = 15, nrow = NULL)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{sample}{A string, representing the unique sample name of interest}

\item{n}{An integer, representing the amount of colors used to depict taxa}

\item{nrow}{An integer, representing the amount of rows in the facet_wrap}
}
\description{
Return a visualization designed for a small number of samples
}
