% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_taco}
\alias{test_taco}
\title{Create a tidytacos object for testing/example purporses}
\usage{
test_taco()
}
\value{
a small tidytacos object
}
\description{
Create a tidytacos object for testing/example purporses
}
