% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{tidy_count_to_matrix}
\alias{tidy_count_to_matrix}
\title{Convert counts tidy data frame to matrix}
\usage{
tidy_count_to_matrix(counts, value = count)
}
\arguments{
\item{counts}{The counts tidy data frame that will be converted.}

\item{value}{Name of column containing the counts data. Default is
"counts".}
}
\description{
\code{tidy_count_to_matrix} returns a numerical matrix given a tidy
counts data frame.
}
\details{
This function will convert a counts tidy data frame into a numerical
counts matrix. To convert a numerical counts matrix into a counts
tidy data frame use \code{\link{tidy_count_to_matrix}}.
}
