% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytacos-package.R
\docType{package}
\name{tidytacos-package}
\alias{tidytacos-package}
\title{tidytacos: Manipulate Taxonomic Composition Data of Microbial Communities}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A toolkit for processing and visualizing taxonomic composition data of microbial communities, based on the tidyverse framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://lebeerlab.github.io/tidytacos/}
  \item Report bugs at \url{https://github.com/LebeerLab/tidytacos/issues/new?assignees=TheOafidian&labels=bug&projects=&template=bug_report.md&title=\%5BBUG\%5D+}
}

}
\author{
\strong{Maintainer}: Stijn Wittouck \email{stijn.wittouck@uantwerpen.be}

Authors:
\itemize{
  \item Tim Van Rillaer \email{tim.vanrillaer@uantwerpen.be}
  \item Wenke Smets \email{wenke.smets@uantwerpen.be}
}

Other contributors:
\itemize{
  \item Sarah Lebeer \email{sarah.lebeer@uantwerpen.be} [funder]
}

}
\keyword{internal}
