% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{to_biom}
\alias{to_biom}
\title{Write the counts of the tidytacos object to a biom file}
\usage{
to_biom(ta, filename = "asvs.biom")
}
\arguments{
\item{ta}{A tidytacos object.}

\item{filename}{The name of the resulting biom table file, defaults to 'asvs.biom'.}
}
\description{
Uses the taxon_id and sample_id columns to create a dense biom file (v1, json). 
By default this is on ASV/OTU level. 
To do it at any other taxonomy level, one first needs to aggregate the taxa.
}
