% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{trim_asvs}
\alias{trim_asvs}
\title{Trim all sequences}
\usage{
trim_asvs(ta, start, end)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{start}{Index of where to start trimming.}

\item{end}{Index of where to stop trimming.}
}
\value{
A tidytacos object.
}
\description{
\code{trim_asvs} trims sequence ends of the sequence supplied in the taxa table. 
This function assumes that the sequence variable in the taxon table is called
"sequence".
}
\examples{
# keep only the first 200 nucleotides of the sequences
urt \%>\% trim_asvs(0, 200)
}
