% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{write_tidytacos}
\alias{write_tidytacos}
\title{Write community data in tidytacos format}
\usage{
write_tidytacos(ta, dout)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{dout}{The directory to store the three tidytacos tables in.}
}
\description{
`write_tidytacos()` saves the tidytacos object into 3 .csv files. This format allows easy loading of the tidytacos object using the [read_tidytacos()] function.
}
