% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_abundances.R
\name{add_absolute_abundance}
\alias{add_absolute_abundance}
\title{Add absolute abundances to count table}
\usage{
add_absolute_abundance(ta, spike_taxon, spike_added = spike_added)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{spike_taxon}{The taxon id of the spike.}

\item{spike_added}{The column name in the sample table which indicates how
much spike was added per sample.}
}
\value{
A tidytacos object with absolute abundances added to the count table.
}
\description{
\code{add_absolute_abundance()} calculates absolute abundances of taxa in
samples given a taxon that was spiked into all of the samples during library
prep. The function then adds these absolute abundances to the count table of
the tidytacos object under the column name "absolute_abundance".
}
\examples{
# Initiate count matrix
x <- matrix(
  c(1500, 1300, 14, 280, 356, 9),
  ncol = 2
)
rownames(x) <- c("taxon1", "taxon2", "taxon3")
colnames(x) <- c("sample1", "sample2")

# Convert to tidytacos object
data <- create_tidytacos(x,
  taxa_are_columns = FALSE
)
data$samples$spike_added <- c(100, 150)

# Add total abundance
data <- data \%>\%
  add_absolute_abundance(spike_taxon = "t3")
}
\seealso{
Other count-modifiers: 
\code{\link{add_density}()},
\code{\link{add_rel_abundance}()}
}
\concept{count-modifiers}
