% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_samples.R
\name{add_alpha}
\alias{add_alpha}
\title{Add alpha diversity measure}
\usage{
add_alpha(
  ta,
  method = "invsimpson",
  keep_empty_samples = FALSE,
  subsample = FALSE,
  ...
)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{method}{The diversity measure to use,
see \code{\link[vegan:diversity]{vegan::diversity()}} for further information on these.}

\item{keep_empty_samples}{Whether to keep empty samples or not.}

\item{subsample}{Wether to use subsampling to be able to compare samples of varying sequencing depths.
\href{https://journals.asm.org/doi/10.1128/msphere.00355-23}{Schloss et al., 2023}}

\item{...}{
  Arguments passed on to \code{\link[=add_subsampled_alpha]{add_subsampled_alpha}}
  \describe{
    \item{\code{min_lib_size}}{the minimum lib size samples need to have.
Samples with lower lib sizes will be discarded and samples with a higher readcount will be itteratively
subsampled to this readcount to allow for a fair comparison across read_depths.}
    \item{\code{itterations}}{the amount of itterations for subsampling. Please report this number in your research.}
  }}
}
\value{
A tidytacos object with the alpha diversity measure added.
}
\description{
\code{add_alpha()} adds an alpha diversity measures to the sample table of a
tidytacos object.
}
\details{
This function can add different alpha diversity measures
to the sample table, specified by the method argument.
The following methods are available:
\itemize{
\item invsimpson: Inverse Simpson index
\item shannon: Shannon index
\item simpson: Simpson index
\item pielou: Pielou's evenness index
\item obs: Observed richness
\item s.chao1: Chao1 richness estimator
\item s.ace: ACE richness estimator
}
}
\examples{
# Initiate counts matrix
x <- matrix(
  c(1500, 1300, 280, 356),
  ncol = 2
)
rownames(x) <- c("taxon1", "taxon2")
colnames(x) <- c("sample1", "sample2")

# Convert to tidytacos object
data <- create_tidytacos(x,
  taxa_are_columns = FALSE
)

# Add alpha diversity measures
data <- data \%>\%
  add_alpha()

data <- data \%>\%
  add_alpha(method = "shannon")
}
\seealso{
Other sample-modifiers: 
\code{\link{add_alphas}()},
\code{\link{add_dominant_taxa}()},
\code{\link{add_metadata}()},
\code{\link{add_ord}()},
\code{\link{add_sample_clustered}()},
\code{\link{add_spike_ratio}()},
\code{\link{add_subsampled_alpha}()},
\code{\link{add_total_absolute_abundance}()},
\code{\link{add_total_count}()},
\code{\link{add_total_density}()},
\code{\link{cluster_samples}()}

Other diversity-metrics: 
\code{\link{add_alphas}()},
\code{\link{add_ord}()},
\code{\link{add_subsampled_alpha}()}
}
\concept{diversity-metrics}
\concept{sample-modifiers}
