% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_samples.R
\name{add_alpha}
\alias{add_alpha}
\title{Add alpha diversity measure}
\usage{
add_alpha(ta, method = "invsimpson")
}
\arguments{
\item{ta}{A tidytacos object.}

\item{method}{The diversity measure to use, see \code{\link[vegan]{diversity}} for further information on these.}
}
\value{
A tidytacos object with the alpha diversity measure added.
}
\description{
`add_alpha()` adds an alpha diversity measures to the sample table of a
tidytacos object.
}
\details{
This function can add different alpha diversity measures to the sample table, specified by the method argument.
The following methods are available:
- invsimpson: Inverse Simpson index
- shannon: Shannon index
- simpson: Simpson index
- pielou: Pielou's evenness index
- obs: Observed richness
- s.chao1: Chao1 richness estimator
- s.ace: ACE richness estimator
}
\examples{
# Initiate counts matrix
x <- matrix(
 c(1500, 1300, 280, 356),
 ncol = 2
)
rownames(x) <- c("taxon1", "taxon2")
colnames(x) <- c("sample1", "sample2")

# Convert to tidytacos object
data <- create_tidytacos(x,
                     taxa_are_columns = FALSE
                     )

# Add alpha diversity measures
data <- data \%>\%
 add_alpha()

data <- data \%>\% 
 add_alpha(method="shannon")

}
