% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_samples.R
\name{add_alphas}
\alias{add_alphas}
\title{Add alpha diversity measures}
\usage{
add_alphas(ta, methods = "all")
}
\arguments{
\item{ta}{A tidytacos object.}

\item{methods}{A character vector of the diversity measure to use, see \code{\link[=add_alpha]{add_alpha()}} for examples.
Optionally use 'all' to add all diversity measures.}
}
\value{
A tidytacos object with the selected alpha diversity measures added.
}
\description{
\code{add_alpha()} adds selected alpha diversity measures to the sample table of a
tidytacos object.
}
\details{
This function can add multiple different alpha diversity measures to the sample table, specified by the methods argument.
}
\examples{
urt_all_alphas <- urt \%>\% add_alphas()
}
\seealso{
Other sample-modifiers: 
\code{\link{add_alpha}()},
\code{\link{add_metadata}()},
\code{\link{add_ord}()},
\code{\link{add_sample_clustered}()},
\code{\link{add_spike_ratio}()},
\code{\link{add_total_absolute_abundance}()},
\code{\link{add_total_count}()},
\code{\link{add_total_density}()},
\code{\link{cluster_samples}()}

Other diversity-metrics: 
\code{\link{add_alpha}()},
\code{\link{add_ord}()}
}
\concept{diversity-metrics}
\concept{sample-modifiers}
