% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_samples.R
\name{add_alphas}
\alias{add_alphas}
\title{Add alpha diversity measures}
\usage{
add_alphas(ta, methods = "all")
}
\arguments{
\item{ta}{A tidytacos object.}

\item{methods}{A character vector of the diversity measure to use, see [add_alpha()] for examples.
Optionally use 'all' to add all diversity measures.}
}
\value{
A tidytacos object with the selected alpha diversity measures added.
}
\description{
`add_alpha()` adds selected alpha diversity measures to the sample table of a
tidytacos object.
}
\details{
This function can add multiple different alpha diversity measures to the sample table, specified by the methods argument.
}
\examples{
urt_all_alphas <- urt \%>\% add_alphas()
}
