% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{add_clr_abundance}
\alias{add_clr_abundance}
\title{Perform a centered log ratio transformation on the readcounts.}
\usage{
add_clr_abundance(ta, overwrite = F)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{overwrite}{Whether or not the counts table is to be overwritten with the transformed counts.}
}
\value{
A tidytacos object.
}
\description{
`add_clr_abundance()` calculates the log ration transformed values for each taxon in each sample and adds these data in a new table, clr_counts. Alternatively, using 'overwrite', the clr transformed data can replace the 'counts' column in the count table.
}
