% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_abundances.R
\name{add_density}
\alias{add_density}
\title{Add density to counts table}
\usage{
add_density(
  ta,
  spike_taxon,
  spike_added = spike_added,
  material_sampled = material_sampled
)
}
\arguments{
\item{ta}{tidytacos object.}

\item{spike_taxon}{The taxon id of the spike.}

\item{spike_added}{The column name of the samples table which indicates how much spike was added per sample, e.g. 16S rRNA gene copy numbers added to the DNA extraction tube.}

\item{material_sampled}{The column name indicating the amount of material from which DNA was extracted, e.g gram of soil. This parameter encourages researchers to consider that absolute abundances are only meaningful if they can be translated into densities.}
}
\description{
\code{add_density} adds density to the counts table of a
tidytacos object.
}
\details{
This function adds the density per taxon and per sample to the counts table
of a tidytacos object under the variable name "density".
}
\examples{
# Initiate count matrix
x <- matrix(
  c(1500, 1300, 14, 280, 356, 9),
  ncol = 2
)
rownames(x) <- c("taxon1", "taxon2", "taxon3")
colnames(x) <- c("sample1", "sample2")

# Convert to tidytacos object
data <- create_tidytacos(x,
  taxa_are_columns = FALSE
)
data$samples$spike_added <- c(100, 150)
data$samples$grams_source <- c(4, 5)

# Add density
data <- data \%>\%
  add_density(spike_taxon = "t3", material_sampled=grams_source)

}
