% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_samples.R
\name{add_metadata}
\alias{add_metadata}
\title{Add metadata to the tidytacos object}
\usage{
add_metadata(ta, metadata, table_type = "sample")
}
\arguments{
\item{ta}{A tidytacos object.}

\item{metadata}{A tibble containing data for each sample or taxon.
Samples/taxa should be rows, while metadata variables should be columns. At
least one column name needs to be shared with the sample or taxa table of
the tidytacos object. The default shared column name is 'sample' for
samples and 'taxon' for taxa.}

\item{table_type}{The type of table to add, either 'sample' or 'taxa'.}
}
\description{
\code{add_metadata} adds sample or taxon metadata to the sample or taxon
table, respectively, of a tidytacos object.
}
\examples{
# Initiate counts matrix
x <- matrix(
 c(1500, 1300, 280, 356),
 ncol = 2
)
rownames(x) <- c("taxon1", "taxon2")
colnames(x) <- c("sample1", "sample2")

# Convert to tidytacos object
data <- create_tidytacos(x,
                     taxa_are_columns = FALSE
                     )

# Initiate sample tibble
sample <- c("sample1", "sample2")
environment <- c("food fermentation", "human stool")
sample_tibble <- tibble::tibble(sample, environment)

# Add sample tibble to tidytacos object
data <- data \%>\%
add_metadata(sample_tibble)

}
