% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_samples.R
\name{add_ord}
\alias{add_ord}
\title{Add ordination}
\usage{
add_ord(ta, distance = "bray", method = "pcoa", dims = 2, binary = FALSE, ...)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{distance}{The distance indices to use, see
\code{\link[vegan]{vegdist}}.}

\item{method}{The ordination method to use to calculate coordinates. Choices
are "pcoa", "tsne", "umap".}

\item{dims}{The amount of dimensions to reduce the dissimilarities to.}

\item{binary}{Perform presence/absence standardisation before distance
computation or not.}

\item{...}{Additional arguments to pass to the ordination function.}
}
\value{
A tidytacos object with the ordination coordinates added.
}
\description{
\code{add_ord} adds the first n dimensions of a dimensionality reduction
method performed on a given dissimilarity matrix as new variables to the
sample table of a tidytacos object.
}
\details{
This function calculates the dissimilarities between samples followed by
an ordination analysis. It will then add the first n dimensions to
the sample table of a tidytacos object named "ord1", "ord2", ... This
function will also add relative abundances if not present using
\code{\link{add_rel_abundance}}.
}
\examples{
# Initiate counts matrix
x <- matrix(
 c(1500, 1300, 280, 356, 456, 678),
 ncol = 3
)
rownames(x) <- c("taxon1", "taxon2")
colnames(x) <- c("sample1", "sample2", "sample3")

# Convert to tidytacos object
data <- create_tidytacos(x,
                     taxa_are_columns = FALSE
                     )

# Add pcoa
data <- data \%>\%
 add_ord()

# The variances of the ordination dimensions can be accessed with
data$ord_variances
}
