% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_samples.R
\name{add_ord}
\alias{add_ord}
\title{Add dimensionality ordination}
\usage{
add_ord(ta, distance = "bray", method = "pcoa", dims = 2, binary = FALSE, ...)
}
\arguments{
\item{ta}{tidytacos object.}

\item{distance}{the distance indices to use, see \code{\link[vegan]{vegdist}}}

\item{method}{the ordination method to use to calculate coordinates. Choice from pcoa, tsne, umap}

\item{dims}{the amount of dimensions to reduce the distances to.}

\item{binary}{perform presence/absence standardisation before distance computation.}
}
\description{
\code{add_ord} adds the first x dimensions of a dimensionality reduction method
of a given dissimilarity matrix to two new variables of the
samples tibble of a tidytacos object.
}
\details{
This function calculates the distance between samples followed by
an ordination analysis. It will then add the first n dimensions to
the samples tibble of a tidytacos object named "ord1", "ord2", ... This
function will also add relative abundances if not present using
\code{\link{add_rel_abundance}}.
}
\examples{
# Initiate counts matrix
x <- matrix(
 c(1500, 1300, 280, 356, 456, 678),
 ncol = 3
)
rownames(x) <- c("taxon1", "taxon2")
colnames(x) <- c("sample1", "sample2", "sample3")

# Convert to tidytacos object
data <- create_tidytacos(x,
                     taxa_are_columns = FALSE
                     )

# Add pcoa
data <- data \%>\%
 add_ord()

}
