% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_taxa.R
\name{add_prevalence}
\alias{add_prevalence}
\title{Add taxon prevalences to the taxon table}
\usage{
add_prevalence(ta, condition = NULL, relative = F, fischer_test = F)
}
\arguments{
\item{ta}{A tidytacos object}

\item{condition}{A categorical variable (string)}

\item{relative}{wether to use relative occurences}

\item{fischer_test}{wether to perform a fischer test and add the p-values of the test to the taxa table}
}
\description{
Adds taxon prevalences (overall or per condition) to the taxa table.
}
\details{
Condition should be a categorical variable present in the samples table.
Supply condition as a string.
}
