% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_samples.R
\name{add_sample_clustered}
\alias{add_sample_clustered}
\title{Add clustering-based sample order}
\usage{
add_sample_clustered(ta)
}
\arguments{
\item{ta}{A tidytacos object.}
}
\description{
`add_sample_clustered()` adds a new variable defining a sample order based
on a hierarchical clustering of the samples.
}
\details{
This function calculates the Bray-Curtis distances between samples followed
by hierarchical average linkage clustering of samples. It will then add a new
factor variable "sample_clustered" to the sample tibble of a tidytacos
object. This function is useful if one wants to plot similar samples
together.
}
\examples{
urtc <- urt \%>\% add_sample_clustered()
urtc$samples \%>\% dplyr::select(sample_id, sample_clustered)
}
