% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_samples.R
\name{add_total_absolute_abundance}
\alias{add_total_absolute_abundance}
\title{Add total absolute abundances of samples}
\usage{
add_total_absolute_abundance(ta, spike_taxon, spike_added = spike_added)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{spike_taxon}{The taxon id of the spike.}

\item{spike_added}{The column name of the samples table which indicates how
much spike was added per sample, e.g. 16S rRNA gene copy numbers added to
the DNA extraction tube.}
}
\value{
A tidytacos object with
the total absolute abundances added to the sample table.
}
\description{
\code{add_total_absolute_abundance()} calculates the total absolute abundances
of the samples given a spike taxon, and adds this to the sample table under
the column name "total_absolute_abundance".
}
\examples{
# Initiate count matrix
x <- matrix(
  c(1500, 1300, 14, 280, 356, 9),
  ncol = 2
)
rownames(x) <- c("taxon1", "taxon2", "taxon3")
colnames(x) <- c("sample1", "sample2")

# Convert to tidytacos object
data <- create_tidytacos(x,
  taxa_are_columns = FALSE
)
data$samples$spike_added <- c(100, 150)

# Add total abundance
data <- data \%>\%
  add_total_absolute_abundance(spike_taxon = "t3")
}
\seealso{
Other sample-modifiers: 
\code{\link{add_alpha}()},
\code{\link{add_alphas}()},
\code{\link{add_metadata}()},
\code{\link{add_ord}()},
\code{\link{add_sample_clustered}()},
\code{\link{add_spike_ratio}()},
\code{\link{add_total_count}()},
\code{\link{add_total_density}()},
\code{\link{cluster_samples}()}
}
\concept{sample-modifiers}
