% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_samples.R
\name{calculate_unifrac_distances}
\alias{calculate_unifrac_distances}
\title{Calculate unifrac distance matrix from a tidytacos object with a rooted tree}
\usage{
calculate_unifrac_distances(ta, ...)
}
\arguments{
\item{ta}{A tidytacos object with a rooted tree in the "tree" slot.}

\item{...}{Additional arguments to pass to the [phyloseq::UniFrac()] function.}
}
\description{
Calculate unifrac distance matrix from a tidytacos object with a rooted tree
}
