% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_samples.R
\name{cluster_samples}
\alias{cluster_samples}
\title{Clusters samples into n clusters}
\usage{
cluster_samples(ta, n_clusters)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{n_clusters}{The number of desired clusters.}
}
\value{
A tidytacos object with a cluster column in the samples table.
}
\description{
\code{cluster_samples()} clusters the samples into n clusters and adds these
clusters to a new variable "cluster" in the sample table.
}
\details{
This function calculates the Bray-Curtis distance between samples followed by
hierarchical average linkage clustering of samples. The user provides a
number of desired clusters which will be used to assign the samples to. A new
variable named "cluster" will be added to the samples tibble of a
tidytacos object defining to what cluster a sample belongs.
}
\examples{
# Initiate count matrix
x <- matrix(
  c(1500, 1300, 280, 356),
  ncol = 2
)
rownames(x) <- c("taxon1", "taxon2")
colnames(x) <- c("sample1", "sample2")

# Convert to tidytacos object
data <- create_tidytacos(x,
  taxa_are_columns = FALSE
)

# Add total abundance
data <- data \%>\%
  cluster_samples(n_clusters = 2)

}
\seealso{
Other sample-modifiers: 
\code{\link{add_alpha}()},
\code{\link{add_alphas}()},
\code{\link{add_dominant_taxa}()},
\code{\link{add_metadata}()},
\code{\link{add_ord}()},
\code{\link{add_sample_clustered}()},
\code{\link{add_spike_ratio}()},
\code{\link{add_subsampled_alpha}()},
\code{\link{add_total_absolute_abundance}()},
\code{\link{add_total_count}()},
\code{\link{add_total_density}()}
}
\concept{sample-modifiers}
