% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{filter_counts}
\alias{filter_counts}
\title{Filter the counts}
\usage{
filter_counts(ta, ...)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{...}{Filter criteria for the counts table.}
}
\value{
A tidytacos object.
# remove singletons
urt <- urt %>% filter_counts(count > 1)
}
\description{
Filter the counts
}
