% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{from_dada}
\alias{from_dada}
\title{DADA2 to a tidytacos object}
\usage{
from_dada(seqtab, taxa, taxa_are_columns = FALSE)
}
\arguments{
\item{seqtab}{Sequence table, output of dada2::makeSequenceTable.}

\item{taxa}{Taxa table, output of dada2::assignTaxonomy.}
}
\description{
\code{from_dada} returns a tidytacos object given a seqtab and taxa object from dada2.
}
\details{
This function will convert two dada2 objects or files into a tidytacos object.
}
