% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{mutate_taxa}
\alias{mutate_taxa}
\title{Create extra variables in the taxa table}
\usage{
mutate_taxa(ta, ...)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{...}{Mutate criteria for the taxa table.}
}
\value{
A tidytacos object.
}
\description{
Create extra variables in the taxa table
}
\examples{
urt <- urt \%>\% mutate_taxa(species = paste(genus, species))
}
