% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adders_samples.R
\name{perform_anosim}
\alias{perform_anosim}
\title{Perform anosim test for statistical difference between groups of samples
Test for a statistical difference between groups of samples
the null hypothesis is that there is no difference between 
microbial communities in the groups of samples}
\usage{
perform_anosim(ta, group, ...)
}
\arguments{
\item{ta}{tidytacos object.}

\item{group}{a column in the samples table to group the samples on.}

\item{distance}{the distance metric to use.}

\item{permutations}{tidytacos object.}
}
\description{
Perform anosim test for statistical difference between groups of samples
Test for a statistical difference between groups of samples
the null hypothesis is that there is no difference between 
microbial communities in the groups of samples
}
