% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{rarefy}
\alias{rarefy}
\title{Rarefy the samples to a given number of reads}
\usage{
rarefy(ta, n, replace = F)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{n}{Subsample size for rarefying the community.}

\item{replace}{Whether to replace the read after it has been selected for the subsample so it can be sampled again. Default is FALSE.}
}
\value{
A tidytacos object.
}
\description{
This function performs rarefying. Make sure that all samples contain at least
the minimum number of reads; otherwise, an error might be thrown.
}
\examples{
# discard samples with less than 1000 reads
urt_1000 <- urt \%>\%
  add_total_count() \%>\%
  filter_samples(total_count >= 1000)

# then rarefy to 1000 reads
urt_1000 <- urt_1000 \%>\% rarefy(1000)

}
