% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{read_tidytacos}
\alias{read_tidytacos}
\title{Read community data written by tidytacos}
\usage{
read_tidytacos(
  din,
  samples = "samples.csv",
  taxa = "taxa.csv",
  counts = "counts.csv"
)
}
\arguments{
\item{din}{directory containing
the sample, taxa and counts table in csv format}

\item{samples}{the name of the samples table, defaults to samples.csv}

\item{taxa}{the name of the taxa table, defaults to taxa.csv}

\item{counts}{the name of the counts table, defaults to counts.csv}
}
\value{
A tidytacos object.
}
\description{
\code{\link[=read_tidytacos]{read_tidytacos()}} reads the three .csv files created by
the \code{\link[=write_tidytacos]{write_tidytacos()}} function and returns a tidytacos object.
}
\details{
The samples.csv file should contain a column named "sample_id".
The taxa.csv file should contain a column named "taxon_id" and at the very least one rank name.
The default rank names used by tidytacos are "domain", "phylum", "class", "order", "family", "genus" and "species".
The counts.csv file should contain columns named "sample_id", "taxon_id" and "count".
}
\seealso{
Other import-methods: 
\code{\link{create_tidytacos}()},
\code{\link{from_dada}()},
\code{\link{from_phyloseq}()}
}
\concept{import-methods}
