% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{select_taxa}
\alias{select_taxa}
\title{Retain or remove a set of taxon variables}
\usage{
select_taxa(ta, ...)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{...}{Selection criteria for the taxa table.}
}
\value{
A tidytacos object.
}
\description{
Retain or remove a set of taxon variables
}
\examples{

# drop the sequence column
urt \%>\% select_taxa(-sequence)

# keep only the taxon_id and genus columns
urt \%>\% select_taxa(taxon_id, genus)

}
