% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{tacoplot_ord}
\alias{tacoplot_ord}
\title{Return an ordination plot of the samples}
\usage{
tacoplot_ord(
  ta,
  x = NULL,
  palette = NULL,
  ord = "pcoa",
  distance = "bray",
  stat.method = NULL,
  title = NULL,
  ...
)
}
\arguments{
\item{ta}{A tidytacos object.}

\item{x}{The column name used to color the sample groups on.}

\item{palette}{A vector of colors, used as the palette for coloring sample
groups.}

\item{ord}{the ordination technique to use. Choice from pcoa, tsne and umap.}

\item{distance}{the distance algorithm to use, see \code{\link[vegan]{vegdist}}.}

\item{stat.method}{the statistic to print on the figure, choice from mantel and anosim.}

\item{title}{a string to display as title of the plot.}

\item{...}{Extra arguments to pass to the add_ord function.}
}
\description{
Creates an ordination plot of the beta diversity of the samples in the tidytacos object.
This can be used to gauge the similarity between samples.
}
\examples{

tacoplot_ord(urt, x=location)

# set plate to character, to avoid it being treated as a continuous variable
urt <- urt \%>\% mutate_samples(plate=as.character(plate))
tacoplot_ord(urt, x=plate, ord="umap", distance="aitchison",stat.method="permanova")

}
